comment updates Database.java
I updated/added javadoc formatted comments to each method, the class, and each of the class variables.

One thing to note is that there is a random variable in here called u. I am unsure what it is and it doesn't seem to be used for anything but I am not 100% sure as I didn't read all of the code through to look for if it is used other places in this method besides as a variable and in the constructor. We might be able to just remove it. There is a previous comment there for it that just says "Just testing stuff...Relax" so I am guessing we should be able to remove it