Comment/Organization updates Report.java
Organized all of the private helper methods to be in the order as they are listed in the createObjects() method (which is also the same order as the ArrayList variables are declared in the beginning of the file). Note: I didn't modify any of their code, just rearranged the order the methods are in to make it match the pattern

Added javadoc comments for all the ArrayList variables

Fixed, added descriptions, and formatted all of the javadoc comments for the helper methods (eg. fetchFMS())

Overall, ensured that all javadoc comments were properly formatted and had descriptions as well as descriptions for the @param and @return tags