Cleaned up BowlerController.java
As this file is over 3000 lines long, I went through to remove any unnecessary blank lines. And by unnecessary, I left many in, just in sections where it obviously isnt necessary for any sort of formatting or to make it easier to read I went ahead and removed those blank lines. Additionally I went ahead and cleaned up some of the comments and their weird formatting.

One thing that we should eventually do to clean up the code in this file is to look at the different regular expression variables inside of TextFieldRequired.java to use those from lines 877 - 921 where applicable.

Fixed some spots in the code where it was obvious that a previous group copy and pasted lines repeatedly without properly updating all of the comments