Comment/organization Updates AdminController.java
Added javadoc comments to the variables as well as the methods that didnt have one. Updated the descriptions in most to make them more descriptive. Rearranged the showNewAthlete, showSearch, manageQuestions, and showForms methods to get it to match the application and how the buttons appear on the page. 

Cleared out 14 blank lines from below the last closing bracket and 2 from immediately above that